% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_fitting_call.R
\name{construct_fitting_call}
\alias{construct_fitting_call}
\title{Construct a call for refitting a model from the model itself}
\usage{
construct_fitting_call(model, data_name = "training", ...)
}
\arguments{
\item{model}{the model in question}

\item{data_name}{character string specifying the name of the data
frame used for the refitting. This object \emph{must} be defined in the environment in which the
call is being made.}

\item{...}{(not used)}
}
\description{
This will typically \emph{not} be used by the end-user.
}
\details{
This provides a way to refit a model on either resampled or sub-sampled data.
Not all model architectures support this. If not, then you can't use \code{mod_ensemble} or \code{mod_cv},
or use the \code{bootstrap=} argument in any of the other functions.
}
