% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_statistics.R
\name{coverage}
\alias{coverage}
\title{Interval statistics for use with df_stats()}
\usage{
coverage(level = 0.95)
}
\arguments{
\item{level}{Confidence level, e.g. 0.95}
}
\description{
Interval statistics for use with df_stats()
}
