% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counts.R
\name{df_counts}
\alias{df_counts}
\title{Formula interface to counts}
\usage{
df_counts(formula, data, wide = FALSE, margins = FALSE)
}
\arguments{
\item{formula}{the formula describing the relationship}

\item{data}{a data frame (or you can pipe this in)}

\item{wide}{reformat the output as a cross-tabulation. This makes sense only when there are just two variables}

\item{margins}{show the marginal counts. Makes the most sense if \code{wide = TRUE}.}
}
\description{
Counts the number of cases in a data frame broken down by the
variables in the formula.
}
\seealso{
\code{df_props}
}
