% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explanatory_vars.R
\name{explanatory_vars}
\alias{explanatory_vars}
\title{Get the names of the explanatory variables in a model}
\usage{
explanatory_vars(model, ...)
}
\arguments{
\item{model}{the model in question}

\item{...}{(not used)}
}
\description{
This will typically \emph{not} be used by the end_user.
}
