% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_from_mod.R
\name{formula_from_mod}
\alias{formula_from_mod}
\title{Extract the model formula used in specifying the model}
\usage{
formula_from_mod(model, ...)
}
\arguments{
\item{model}{the model}

\item{...}{(not used)}
}
\description{
This typically will \emph{not} be used by an end-user.
}
\details{
Not all model architectures support this. If a model architecture isn't recognized,
you'll have to add a method for that class. See vignette.
}
