% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_statistics.R
\name{coverage}
\alias{coverage}
\alias{ci.mean}
\alias{ci.median}
\alias{ci.sd}
\alias{ci.proportion}
\title{Interval statistics}
\usage{
coverage(x, level = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{A variable.}

\item{level}{Number in 0 to 1 specifying the confidence level for the interval. (Default: 0.95)}

\item{na.rm}{if \code{TRUE} disregard missing data}

\item{prop_of}{For proportions, this specifies the categorical level for which the calculation of proportion will
be done. Defaults: \code{TRUE} for logicals for which the proportion is to
be calculated.}
}
\description{
Calculate coverage intervals and confidence intervals for the sample mean, median, sd, proportion, ...
}
\details{
Typically, these will be used within \code{df_stats()}. For the mean, median, and sd, the variable x must be
quantitative. For proportions, the x can be anything; use the \code{prop_of} argument to specify what
value you want the proportion of. Default for \code{prop_of} is \code{TRUE} for x logical, or the first level returned
by \code{unique} for categorical or numerical variables.  The CI for proportions is computed using quantiles of the
binomial distribution.
}
\examples{
# The central 95\% interval
df_stats(hp ~ cyl, data = mtcars, c95 = coverage(0.95))
# The confidence interval on the mean
df_stats(hp ~ cyl, data = mtcars, mean, ci.mean)
# What fraction of cars have 6 cylinders?
df_stats(mtcars, ~ cyl, six_cyl_prop = ci.proportion(prop_of = 6, level = 0.90))

}
