% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_architectures.R
\name{mod_architectures}
\alias{mod_architectures}
\title{The model architectures supported in \code{mosaicModel}}
\usage{
mod_architectures()
}
\description{
Identifies the available model architectures so that you can ask for ones that
aren't yet there! Even better, for those with the needed R skills,
if you can create the necessary infrastructure and
put it in a pull request to the repo identified in the \code{DESCRIPTION} file for the package.
}
\details{
The information about the supported architectures is in
an internal object called \code{architectures}. End-users will not need this object. They just have to
know what model architectures are supported.
TO ADD A NEW ARCHITECTURE
\enumerate{
\item Create the appropriate eval_ function
\item Either register the model class and eval_ function in the \code{architectures} table above
or add it to an oversight eval_ fucntion such as eval_glm
\item Create a data_from_model() method for the new class. If it's not possible to extract
the data from the model object, throw an error in your method.
\item Similarly, create methods for explanatory_vars(), response_var(),
response_values(), and formula_from_model()
}
}
