% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_output.R
\name{mod_output}
\alias{mod_output}
\title{Internal functions for evaluating models}
\usage{
mod_output(model, data = NULL, interval = "none", ...)
}
\arguments{
\item{model}{A model object of the classes permitted}

\item{data}{Usually, a data table specifying the inputs to the model. But if
not specified, the training data will be used.}

\item{interval}{One of "none", "confidence", or "prediction". Not all model
types support "prediction" or even "confidence".}

\item{...}{additional arguments}
}
\description{
These functions are the interface to the various model types for \code{mod_eval()}, and through
that to all the other \code{mod_} functions that need to evaluate models, e.g. \code{mod_effect()}, \code{mod_cv()}, and so on.
}
\details{
These functions return a data frame with one numerical vector (for regression types) or
a matrix of probabilities (for classifiers)
}
