% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_train.R
\name{mod_train}
\alias{mod_train}
\title{Training interface for model architectures}
\usage{
mod_train(architecture = NULL, formula, data = NULL, ...)
}
\arguments{
\item{architecture}{A character string specifying the desired architecture.}

\item{formula}{A formula specifying the response and explanatory variables as well as any additional
model terms (such as interactions).}

\item{data}{The dataframe to be used for fitting.}

\item{...}{Other parameters of importance for individual model architectures}
}
\description{
Many model fitting functions in R such as \code{lm}, \code{glm}, and \code{rpart} use a dataframe/formula
interface, where the response and explanatory variables are contained in a single dataframe
and the formula describes what variables and terms are to be included in the model. But some
architectures, e.g. \code{knn}, \code{lda/qda}, etc. do not adopt this convention. This function provides
a way to use the dataframe/formula interface for such architectures.
}
