% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_var.R
\name{response_var}
\alias{response_var}
\title{Get the name or values of the response variable from data used to train a model}
\usage{
response_var(model, ...)
}
\arguments{
\item{model}{the model in question}

\item{...}{(not used)}
}
\description{
This will typically \emph{not} be used by the end-user.
}
