% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_statistics.R
\name{coverage}
\alias{coverage}
\alias{ci.mean}
\alias{ci.median}
\alias{ci.sd}
\title{Interval statistics for use with df_stats()}
\usage{
coverage(level = 0.95)
}
\arguments{
\item{level}{Number in 0 to 1 specifying the confidence level for the interval. (Default: 0.95)}
}
\description{
Function builders for calculating intervals. These must \emph{always} be evaluated
with the \emph{result} being handed as a argument to \code{df_stats()}.
}
\examples{
cover <- coverage(0.95)
df_stats(hp ~ cyl, data = mtcars, c95 = cover)

}
