% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_train.R
\name{mod_train}
\alias{mod_train}
\title{Training statistical models}
\usage{
mod_train(data, formula = formula(data), architecture = "lm")
}
\arguments{
\item{data}{the data to use for training}

\item{formula}{the formula describing the structure of the relationship among variables}

\item{architecture}{the model architecture, given as a function name, quoted or not, e.g. lm or "lm"}
}
\description{
This function coordinates four components in building statistical models:
\enumerate{
\item the data to use for training
\item the response variable and
\item the explanatory variables, specified as a model formula
\item the architecture of the model.
}
}
\details{
This is a light wrapper on the architecture-specific function fitting programs
}
